[OCVL-184] fix property args in sone modes

+++ nodes/objdetect/cv_BRISK.py
@@ -30,7 +30,7 @@ class OCVLBRISKNode(OCVLNodeBase, OCVLFeature2DNode):
 
     thresh_init: bpy.props.IntProperty(default=30, min=5, max=100, update=update_and_init)
     octaves_init: bpy.props.IntProperty(default=3, min=1, max=10, update=update_and_init)
-    patternScale_init: bpy.props.IntProperty(default=1.0, min=0.1, max=10., update=update_and_init)
+    patternScale_init: bpy.props.FloatProperty(default=1.0, min=0.1, max=10., update=update_and_init)
 
     def init(self, context):
         super().init(context)
